model_requirements.txt
pandas_gbq
bigframes
mlflow
google-cloud-aiplatform
google-cloud-bigquery
numpy
scipy
cython
shap
seaborn